requirements.txt
aiohttp==3.5.1
async-timeout==3.0.1
attrs==18.2.0
chardet==3.0.4
discord-rewrite==1.0.0
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
python-dotenv==0.10.1
pyttsx==1.1
typing-extensions==3.6.6
websockets==4.0.1
yarl==1.3.0
